+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/config/constants/MonitorConfig.java
@@ -321,7 +321,7 @@ public class MonitorConfig {
   public static final String SAMPLE_PARTITION_METRICS_STORE_DURING_EXECUTION_CLASS_CONFIG = "sample.partition.metrics.store.during.execution.class";
   public static final String DEFAULT_SAMPLE_PARTITION_METRICS_STORE_DURING_EXECUTION_CLASS = NoopSampleStore.class.getName();
   public static final String SAMPLE_PARTITION_METRICS_STORE_DURING_EXECUTION_CLASS_DOC = "The sample store during execution class name. "
-      + "User may configure a sample store that persists the partition metric samples collected while there is an ongoing execution."
+      + "User may configure a sample store that persists the partition metric samples collected while there is an ongoing execution. "
       + "This config is not intended for enhancing the fault tolerance of the system through recovery of the historical load information.";
 
   /**
