+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/monitor/sampling/ReadOnlyKafkaSampleStore.java
@@ -36,7 +36,6 @@ public class ReadOnlyKafkaSampleStore extends KafkaSampleStore {
     for (int i = 0; i < numProcessingThreads; i++) {
       _consumers.add(createSampleStoreConsumer(config, CONSUMER_CLIENT_ID_PREFIX));
     }
-    _loadingProgress = -1.0;
   }
 
   @Override
